Rasterize channel: fix broadcasting error
In some cases, the at the end of the process (when merging rasters), the error "ValueError: could not broadcast input array from shape (1000,1000) into shape (1000,)" would occur. This has been fixed now.

The root cause seems to be a small bug in the QGIS native Rasterize Mesh algorithm: output raster resolution is not exactly equal to the requested resolution