fixed offset_curve_fixed function and updated to new 3Di database schema
1. updated to new 3Di database schema
2. updated the function 'offset_curve_fixed':

it sometimes happen that a very small channel part, i.e. 1x10^-16 is given to the offset_curve_fixed function, by the fill_wedge function called at line 1371 channel1.fill_wedge(channel2). it is real-world-wise not possible, and it gives an "shapely.errors.GEOSException: vector too long" error.
this is fixed by setting the margin of distance == 0 to -1 to 1 mm channel length